Mostly works

+++ edit.go
@@ -9,8 +9,55 @@ import (
 	"os/exec"
 	"path"
 	"path/filepath"
+	"sort"
+	"strings"
 	"time"
 )
+
+func parseEntries(text string) []string {
+	entry := ""
+	entries := make(map[int]string)
+	var gt int
+	for _, line := range strings.Split(text, "\n") {
+		entryWords := strings.Split(strings.TrimSpace(line), " ")
+		if len(entryWords) > 1 {
+			t1, e1 := time.Parse("2006-01-02 15:04:05", entryWords[0]+" "+entryWords[1])
+			t2, e2 := time.Parse("2006-01-02 15:04", entryWords[0]+" "+entryWords[1])
+			if e1 == nil || e2 == nil {
+				if len(entry) > 0 {
+					if _, ok := entries[gt]; ok {
+						logger.Warn("Duplicate entry for %s", entryWords[0]+" "+entryWords[1])
+					}
+					entries[gt] = strings.TrimSpace(entry)
+				}
+				entry = ""
+				if e1 == nil {
+					gt = int(t1.Unix())
+				} else {
+					gt = int(t2.Unix())
+				}
+			}
+		}
+		entry += strings.TrimRight(line, " ") + "\n"
+	}
+	if len(entry) > 0 {
+		entries[gt] = strings.TrimSpace(entry)
+	}
+
+	// Sort the entries in order
+	var keys []int
+	for k := range entries {
+		keys = append(keys, k)
+	}
+	sort.Ints(keys)
+
+	entriesInOrder := []string{}
+	for _, k := range keys {
+		entriesInOrder = append(entriesInOrder, entries[k])
+	}
+	logger.Debug("Parsed %d entries.", len(entriesInOrder))
+	return entriesInOrder
+}
 
 func cleanUp() error {
 	logger.Debug("Cleaning...")
@@ -51,13 +98,6 @@ com! WPCLI call WordProcessorModeCLI()`), 0644)
 		log.Fatal(err)
 	}
 
-	t := time.Now()
-	dateString := string(t.Format("2006-01-02 15:04:05")) + "   "
-	err = ioutil.WriteFile(path.Join(RuntimeArgs.TempPath, "temp"), []byte(dateString), 0644)
-	if err != nil {
-		log.Fatal(err)
-	}
-
 	cmdArgs := []string{"-u", path.Join(RuntimeArgs.TempPath, "vimrc"), "-c", "WPCLI", "+startinsert", path.Join(RuntimeArgs.TempPath, "temp")}
 	cmd := exec.Command("vim", cmdArgs...)
 	cmd.Stdin = os.Stdin
@@ -78,6 +118,9 @@ func writeEntry(fileContents string, forceWrite bool) {
 	h.Write([]byte(fileContents))
 	sha1_hash := hex.EncodeToString(h.Sum(nil))
 	fileName := string(sha1_hash) + ".gpg"
+	if exists(path.Join(RuntimeArgs.FullPath, fileName)) {
+		return
+	}
 
 	encryptedText := encryptString(string(fileContents), RuntimeArgs.Passphrase)
 	err := ioutil.WriteFile(path.Join(RuntimeArgs.FullPath, fileName), []byte(encryptedText), 0644)
